Show list after filter change
This makes it possible to show the list after filter change. Before this change, the app was not showing list if clicked Show button for some country.

This creates an error or a warning, most likely because I think it's against the rules of hooks to set state inside useEffect.

Internal React error: Expected static flag was missing. Please notify the React team. Error Component Stack
    at Countries (Countries.jsx:5:22)
    at div (<anonymous>)
    at App (App.jsx:7:37)